+++ src/table_new/Ellipsis.tsx
